:truck: move SSE service to event_handling module
